move `tshwrap` to `lib/` from `/tool/tbot/`
