+++ lib/tbot/tshwrap/wrap_test.go
