+++ tool/tbot/db.go
@@ -20,8 +20,8 @@ import (
 	"path/filepath"
 
 	"github.com/gravitational/teleport/lib/tbot/config"
+	"github.com/gravitational/teleport/lib/tbot/tshwrap"
 	"github.com/gravitational/teleport/lib/utils"
-	"github.com/gravitational/teleport/tool/tbot/tshwrap"
 	"github.com/gravitational/trace"
 )
