+++ tool/tbot/proxy.go
@@ -20,7 +20,7 @@ import (
 	"path/filepath"
 
 	"github.com/gravitational/teleport/lib/tbot/config"
-	"github.com/gravitational/teleport/tool/tbot/tshwrap"
+	"github.com/gravitational/teleport/lib/tbot/tshwrap"
 	"github.com/gravitational/trace"
 )
